pyproject.toml
[build-system]
    build-backend = "poetry.core.masonry.api"
    requires      = ["poetry-core>=1.0.0"]
[tool.poetry]
    authors = ["Tim Holzapfel"]
    classifiers = [
        "Environment :: Web Environment :: Mozilla",
        "Development Status :: 5 - Production/Stable",
    ]
    description = "Gather data for the social sciences"
    documentation = "https://seleniumscraping.readthedocs.io/en/latest/index.html"
    exclude = ["src/SeleniumScraping/develop_dir/**"]
    license = "Proprietary"
    name = "SeleniumScraping"
    packages = [{ include = "SeleniumScraping", from = "src" }]
    readme = "README.md"
    repository = "https://github.com/Tim-Lukas-H/SeleniumScraping"
    version = "1.0.0"
[tool.poetry.dependencies]
    black               = "^21.12-beta.0"
    fastparquet         = { version = "^0.7.1", optional = true }
    func-timeout        = "^4.3.5"
    furl                = "^2.1.3"
    geopy               = "^2.2.0"
    importlib-resources = "^5.3.0"
    numexpr             = { version = "^2.7.3", optional = true }
    numpy = "^1.21.5"
    openpyxl            = "^3.0.9"
    pandas = "^1.3.5"
    progressbar2        = "^3.55.0"
    psutil              = "^5.8.0"
    pyarrow             = "^6.0.0"
    pycountry           = "^20.7.3"
    pypiwin32           = "^223"
    pyshp               = "^2.1.3"
    python              = ">=3.7.1,<3.11"
    python-dateutil     = "^2.8.2"
    pytz                = "^2021.3"
    regex               = "^2021.11.10"
    selenium            = "^4.0.0"
    termcolor           = "^1.1.0"
    typing-extensions   = "^4.0.1"
    webdrivermanager    = "^0.10.0"
    wheel               = "^0.37.0"
[tool.poetry.group.dev.dependencies]
    black          = "^21.9-beta.0"
    mypy           = "^0.910"
    pre-commit     = "^2.15.0"
    pycodestyle    = "^2.8.0"
    pydocstyle     = "^6.1.1"
    pyenchant      = "^3.2.2"
    pylint         = "^2.12.1"
    pytest         = "^6.2.5"
    spyder-kernels = "^2.1.3"
[tool.poetry.extras]
    performance = ["numexpr", "fastparquet"]
[tool.poetry.scripts]
    SeleniumScraping = "SeleniumScraping.parser.parsermain:parse_main"
[tool.black]
    line-length = 78
[tool.pycln]
    all = true
[tool.pyright]
    reportUnnecessaryIsInstance = false
